try and fix some bad naming
